backend/app/requirements.txt
uvicorn==0.12.2
fastapi==0.61.1
SQLAlchemy==1.3.20
python-jose==3.2.0
face-recognition==1.3.0
passlib==1.7.4
python-multipart==0.0.5
alembic==1.4.3
pytest==6.1.1
psycopg2==2.8.6
